requirements.txt
appdirs==1.4.4
cfgv==3.2.0
distlib==0.3.1
filelock==3.0.12
identify==1.5.10
importlib-metadata==3.3.0
mysql-connector-python==8.0.22
nodeenv==1.5.0
pre-commit==2.9.3
protobuf==3.14.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
virtualenv==20.2.2
zipp==3.4.0
